chore(core): re-organize rego location within service module
