+++ service/authorization/policies/entitlements/README.md
@@ -1,0 +1,6 @@
+# OpenTDF Platform OPA rego policies
+
+## entitlements.rego
+
+This is the default rego policy that will parse a JWT, 
+traverse the subject mappings, and return the entitlements.
